requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0

sentence-transformers
langchain

chromadb

pypdf

python-dotenv
